product model from accounts to departments
