+++ accounts/admin.py
@@ -2,7 +2,6 @@ from django.contrib import admin
 from .models import *
 
 admin.site.register(Customer)
-admin.site.register(Product)
 admin.site.register(Order)
 admin.site.register(OrderItem)
 admin.site.register(ShippingAddress)
