+++ accounts/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
+from departments.models import Product
 
 
 class Customer(models.Model):
@@ -7,16 +8,6 @@ class Customer(models.Model):
         User, on_delete=models.CASCADE, null=True, blank=True)
     name = models.CharField(max_length=200, null=True)
     email = models.CharField(max_length=200, null=True)
-
-    def __str__(self):
-        return self.name
-
-
-class Product(models.Model):
-    name = models.CharField(max_length=200, null=True)
-    price = models.FloatField(null=True)
-    digital = models.BooleanField(default=False, null=True, blank=False)
-    # image
 
     def __str__(self):
         return self.name
