+++ departments/models.py
@@ -1,3 +1,11 @@
 from django.db import models
 
-# Create your models here.
+
+class Product(models.Model):
+    name = models.CharField(max_length=200, null=True)
+    price = models.FloatField(null=True)
+    digital = models.BooleanField(default=False, null=True, blank=False)
+    # image
+
+    def __str__(self):
+        return self.name
